requirements.txt
pyrogram==2.0.106
tgcrypto==1.2.5
aiohttp==3.9.5
python-dotenv==1.0.1
humanize==4.9.0